Update vip features for treatment
